refactor: move HTTP util stuff from rest to server-common
