+++ rest/src/main/groovy/whelk/rest/api/UserDataAPI.groovy
@@ -5,6 +5,7 @@ import groovy.util.logging.Log4j2 as Log
 import groovy.json.JsonSlurper
 import whelk.Whelk
 import whelk.util.WhelkFactory
+import whelk.util.http.HttpTools
 
 import javax.servlet.http.HttpServlet
 import javax.servlet.http.HttpServletRequest
