+++ server-common/src/main/groovy/whelk/util/http/NotFoundException.groovy
@@ -1,0 +1,7 @@
+package whelk.util.http
+
+class NotFoundException extends NoStackTraceException {
+    NotFoundException(String msg) {
+        super(msg)
+    }
+}
