+++ server-common/src/main/groovy/whelk/util/http/OtherStatusException.groovy
@@ -1,0 +1,10 @@
+package whelk.util.http
+
+class OtherStatusException extends NoStackTraceException {
+    int code
+
+    OtherStatusException(String msg, int code, Throwable cause = null) {
+        super(msg, cause)
+        this.code = code
+    }
+}
