+++ server-common/src/main/groovy/whelk/util/http/RedirectException.groovy
@@ -1,0 +1,7 @@
+package whelk.util.http
+
+class RedirectException extends NoStackTraceException {
+    RedirectException(String msg) {
+        super(msg)
+    }
+}
