+++ server-common/src/main/java/whelk/AuthenticationFilter.java
